% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbartAR.R
\name{cbartAR1}
\alias{cbartAR1}
\title{CBART-GP for time series data.}
\usage{
cbartAR1(
  x.train,
  y.train,
  x.test = matrix(0, 0, 0),
  weights = seq(0, 1, length.out = 6),
  seed = 1
)
}
\arguments{
\item{x.train}{Explanatory variables for training (in sample) data. May be a matrix or a data frame, with (as usual) rows corresponding to observations and columns to variables. If a variable is a factor in a data frame, it is replaced with dummies. Note that q dummies are created if q>2 and one dummy is created if q=2, where q is the number of levels of the factor. %\code{makeind} is used to generate the dummies. \code{wbart} will generate draws of \eqn{f(x)} for each \eqn{x} which is a row of \code{x.train}.}

\item{y.train}{Continuous dependent variable for training (in sample) data.}

\item{x.test}{Explanatory variables for test (out of sample) data. Should have same structure as x.train. \code{wbart} will generate draws of \eqn{f(x)} for each \eqn{x} which is a row of x.test.}

\item{weights}{Weights for searching.}

\item{seed}{seed.}
}
\value{
A list that stores the results of CBART-GP for 1D data.
}
\description{
CBART-GP for time series data.
}
\references{
The paper
\emph{Gaussian processes Correlated Bayesian Additive Regression Trees}.
}
