\name{CbartGP-package}
\alias{CbartGP-package}
\alias{CbartGP}
\docType{package}
\title{\packageTitle{CbartGP}}
\description{\packageDescription{CbartGP}}
\section{Package Content}{\packageIndices{CbartGP}}
\author{\packageAuthor{CbartGP}}
\section{Maintainer}{\packageMaintainer{CbartGP}}
\keyword{package}
