% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbartMatern.R
\name{cbartMatern}
\alias{cbartMatern}
\title{CBART-GP for spatial data.}
\usage{
cbartMatern(
  coords,
  x.train,
  y.train,
  x.test = matrix(0, 0, 0),
  method = "geoR",
  weights = seq(0, 1, length.out = 6),
  seed = 1
)
}
\arguments{
\item{coords}{2D spatial coordinates.}

\item{x.train}{Explanatory variables for training (in sample) data. May be a matrix or a data frame, with (as usual) rows corresponding to observations and columns to variables. If a variable is a factor in a data frame, it is replaced with dummies. Note that q dummies are created if q>2 and one dummy is created if q=2, where q is the number of levels of the factor.}

\item{y.train}{Continuous dependent variable for training (in sample) data.}

\item{x.test}{Explanatory variables for test (out of sample) data. Should have same structure as x.train. \code{wbart} will generate draws of \eqn{f(x)} for each \eqn{x} which is a row of x.test.}

\item{method}{Method for estimating MLE of Matern parameters. The methods "geoR" and "fields" correspond to the related R package names.}

\item{weights}{Weights for searching.}

\item{seed}{seed.}
}
\value{
A list that stores the results of CBART-GP for 2D data.
}
\description{
CBART-GP for spatial data.
}
\references{
The paper
\emph{Correlated Bayesian Additive Regression Trees with Gaussian Processes for Regression Analysis of Dependent Data}.
}
