% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbartAR.R
\name{predCbartAR1}
\alias{predCbartAR1}
\title{Prediction AR1}
\usage{
predCbartAR1(residuals.tr, predCbart.te, xtr, xte, rou)
}
\arguments{
\item{residuals.tr}{The residuals (y-f_CBART) of training data.}

\item{predCbart.te}{The CBART prediction on testing data.}

\item{xtr}{1D positions of training data.}

\item{xte}{1D positions of testing data.}

\item{rou}{The maximum likelihood estimations of AR(1)'s parameter rou.}
}
\value{
A vector of the predicted values of the testing data.
}
\description{
Prediction AR1
}
\references{
The paper
\emph{Correlated Bayesian Additive Regression Trees with Gaussian Processes for Regression Analysis of Dependent Data}.
}
