% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbartMatern.R
\name{predCbartMatern}
\alias{predCbartMatern}
\title{Prediction Matern}
\usage{
predCbartMatern(
  residuals.tr,
  predCbart.te,
  coords.tr,
  coords.te,
  method = NULL,
  matern.cov = NULL
)
}
\arguments{
\item{residuals.tr}{The residuals (y-f_CBART) of training data.}

\item{predCbart.te}{The CBART prediction on testing data.}

\item{coords.tr}{2D spatial coordinates of training data.}

\item{coords.te}{2D spatial coordinates of testing data.}

\item{method}{NULL, "geoR", "fields"}

\item{matern.cov}{The maximum likelihood estimations of matern covariance function's parameters.}
}
\value{
A vector of the predicted values of the testing data.
}
\description{
Prediction Matern
}
\references{
The paper
\emph{Correlated Bayesian Additive Regression Trees with Gaussian Processes for Regression Analysis of Dependent Data}.
}
